repaired test, moved around some test containers
