+++ sqlserver/src/test/resources/container-license-acceptance.txt
@@ -1,1 +1,1 @@
-mcr.microsoft.com/mssql/server:2019-latest
+mcr.microsoft.com/azure-sql-edge:latest
