+++ sqlserver/src/test/scala/zio/sql/sqlserver/SqlServerRunnableSpec.scala
@@ -3,7 +3,7 @@ package zio.sql.sqlserver
 import zio._
 import zio.test._
 import java.util.Properties
-import zio.sql.{ ConnectionPoolConfig, JdbcRunnableSpec, TestContainer }
+import zio.sql.{ ConnectionPoolConfig, JdbcRunnableSpec }
 
 trait SqlServerRunnableSpec extends JdbcRunnableSpec with SqlServerModule {
 
@@ -18,7 +18,7 @@ trait SqlServerRunnableSpec extends JdbcRunnableSpec with SqlServerModule {
 
   val poolConfigLayer = ZLayer.scoped {
     TestContainer
-      .postgres()
+      .sqlServer()
       .map(a =>
         ConnectionPoolConfig(
           url = a.jdbcUrl,
